+++ lib/src/app/views/projet/modifier_projet_view.dart
@@ -4,13 +4,14 @@ import 'package:flutter/material.dart';
 import 'package:jdr_maker/src/app/controllers/navigation_controller.dart';
 import 'package:jdr_maker/src/app/controllers/projet_controller.dart';
 import 'package:jdr_maker/src/app/tools/firebase_global_tool.dart';
-import 'package:jdr_maker/src/app/widgets/boutons/bouton.dart';
+import 'package:jdr_maker/src/app/widgets/boutons/form_bouton.dart';
 import 'package:jdr_maker/src/app/widgets/champs/champ_checkbox.dart';
 import 'package:jdr_maker/src/app/widgets/champs/champ_saisie.dart';
 import 'package:jdr_maker/src/app/widgets/chargement.dart';
 import 'package:jdr_maker/src/app/widgets/entete_application.dart';
 import 'package:jdr_maker/src/app/widgets/interfaces/app_interface/app_interface.dart';
 import 'package:jdr_maker/src/domain/data/couleurs.dart';
+import 'package:jdr_maker/src/domain/enums/form_bouton_type.dart';
 import 'package:jdr_maker/src/domain/models/projet_model.dart';
 import 'package:provider/provider.dart';
 
@@ -47,7 +48,7 @@ class _ModifierProjetViewState extends State<ModifierProjetView> {
 
     setState(() => chargement = true);
     ProjetModel projet = projetController.projet!;
-    projet.nomProjet = nomProjet.text;
+    projet.nom = nomProjet.text;
     projet.isPublic = isPublic;
     await FirebaseGlobalTool.modifierDocument(ProjetModel.nomCollection, projet.id, projet.toMap());
     setState(() => NavigationController.changerView(context, "/options"));
@@ -58,7 +59,7 @@ class _ModifierProjetViewState extends State<ModifierProjetView> {
   @override
   Widget build(BuildContext context) {
     projetController = Provider.of<ProjetController>(context);
-    nomProjet.text = projetController.projet!.nomProjet;
+    nomProjet.text = projetController.projet!.nom;
     isPublic = projetController.projet!.isPublic;
 
     return AppInterface(
@@ -112,23 +113,10 @@ class _ModifierProjetViewState extends State<ModifierProjetView> {
                       ],
                     ),
                   ),
-                  Align(
-                    alignment: Alignment.bottomRight,
-                    child: Bouton(
-                      onTap: modifierProjet,
-                      child: Container(
-                        padding: EdgeInsets.all(20),
-                        decoration: BoxDecoration(
-                          color: Couleurs.violet,
-                          borderRadius: BorderRadius.circular(50),
-                        ),
-                        child: Icon(
-                          Icons.done_rounded,
-                          color: Colors.white,
-                          size: 30,
-                        ),
-                      ),
-                    ),
+                  FormBouton(
+                    boutonType: FormBoutonType.valider,
+                    alignement: Alignment.bottomRight,
+                    action: modifierProjet,
                   ),
                 ],
               ),